+++ changelog.md
@@ -2,6 +2,10 @@
 
 ## [2.7.0-dev]
 
+### enhancements
+
+- keyboard/language selector fixes and improvements (now shows active IME).
+
 ### fix
 
 - discord rich presence not inmediately updated after settings change.